refactor: move data of fonts screen to its own file to make it reusable
